requirements.txt
streamlit
doccano_api_client
requests
json
hashlib
sockets
socket.py
pandas
pybase64
pdfminer
textract
numpy
pillow
webbrowser
pyttsx3
SpeechRecognition
docx2txt
python-docx
antiword

